feat(create): move zksync code to dedicated module
